dxvk(w5.0-fix): ULONG/LONG width + __OSITO_K__ guards in util_env
Quality-review follow-up for W5.0:

- Critical: ULONG was typedef'd to unsigned long (8B on LP64). Windows
  ABI requires 32-bit. Every IUnknown::AddRef/Release returns ULONG, so
  the contract is broken for any future D3D11 caller — wrong field
  width in struct layouts. Same for LONG. Fix: switch both to int/uint.

- Critical: util_env.cpp's getExePath() fell into the __linux__ branch
  on OsitoK (since GCC doesn't predefine anything OS-specific in
  freestanding mode and __OSITO_K__ isn't recognized by the gating).
  That branch readlink()'d /proc/self/exe — which doesn't exist on
  OsitoK, would silently return -ENOENT garbage and propagate into
  Logger and config. Fix: add __OSITO_K__ branch that returns "".

- Critical: util_env.cpp's createDirectory() called std::filesystem on
  the non-Windows path. <filesystem> is heavy and pulls glibc symbols
  not in our stubs. Fix: __OSITO_K__ branch returns false (no-op).

Important issues from review (DEFINE_GUID ODR, QPF freq mismatch,
_Unwind_* return widths, .init_array section) deferred — they're
latent for build-only acceptance and will be addressed in W5.1+ as
the actual code paths get exercised.

+++ arch/x86/lib/dxvk/com_stub/IUnknown.h
@@ -68,7 +68,7 @@ typedef const GUID* REFCLSID;
 #endif
 
 #ifndef ULONG
-typedef unsigned long ULONG;
+typedef unsigned int ULONG;  /* Windows ABI: 32-bit even on LP64 */
 #endif
 #ifndef HRESULT
 typedef int HRESULT;
